# cat-shelter-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cat Shelter API",
        "description": "This is a sample Cat Shelter API server.",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get list of employees",
                "description": "Returns list of employees",
                "operationId": "2058499f556b4bb9477e07e7a0912c43",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employee"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Store new employee",
                "description": "Store new employee and return employee data",
                "operationId": "7683a7b1aec5103190be815e1e30f7da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/EmployeeRequest"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employee"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/employees/{employee}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get employee by ID",
                "description": "Returns employee data",
                "operationId": "eea7041853f3dc89946576969345ac8b",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "description": "ID of the employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employee"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update employee by ID",
                "description": "Updates the employee data",
                "operationId": "0b6501c76361cbc0a64e6d425b5aeef4",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "description": "ID of the employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/EmployeeRequest"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employee"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete employee by ID",
                "description": "Deletes the employee",
                "operationId": "5bb714ea86c5fe57744f2498d3f2c5b8",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "description": "ID of the employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Employee deleted successfully"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Employee": {
                "required": [
                    "name",
                    "position"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "position": {
                        "type": "string"
                    },
                    "cat_shelter_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "EmployeeRequest": {
                "required": [
                    "name",
                    "position"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "position": {
                        "type": "string"
                    },
                    "cat_shelter_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    }
}